feat(internal/diagnose): lint command; break out commands & helpers
