+++ internal/diagnose/command.go
@@ -4,6 +4,14 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var options struct {
+	suites        string
+	listSuites    bool
+	verbose       bool
+	attachmentKey string
+	configFile    string
+}
+
 // Command represents the diagnose command.
 var Command = &cobra.Command{
 	Use:   "diagnose",
